requirements.txt
Flask==3.1.0
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.3
Flask-WTF==1.2.2
Flask-Migrate==4.0.7
SQLAlchemy==2.0.36
pandas==2.2.3
openpyxl==3.1.5
reportlab==4.2.5
pytest==8.3.3
psycopg2-binary==2.9.10
python-dotenv==1.0.1
